% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/krige-contour-lines.R, R/krige-contour.R
\docType{data}
\name{stat_krige_contour_lines}
\alias{stat_krige_contour_lines}
\alias{StatKrigeContourLines}
\alias{geom_krige_contour_lines}
\alias{GeomKrigeContourLines}
\alias{geom_krige_contour}
\alias{stat_krige_contour}
\alias{StatKrigeContour}
\alias{GeomKrigeContour}
\title{Contours of kriging predictions}
\format{
An object of class \code{StatKrigeContourLines} (inherits from \code{Stat}, \code{ggproto}, \code{gg}) of length 4.

An object of class \code{GeomKrigeContourLines} (inherits from \code{GeomPath}, \code{Geom}, \code{ggproto}, \code{gg}) of length 2.

An object of class \code{StatKrigeContour} (inherits from \code{Stat}, \code{ggproto}, \code{gg}) of length 4.

An object of class \code{GeomKrigeContour} (inherits from \code{GeomPolygon}, \code{Geom}, \code{ggproto}, \code{gg}) of length 1.
}
\usage{
stat_krige_contour_lines(
  mapping = NULL,
  data = NULL,
  geom = "krige_contour_lines",
  position = "identity",
  ...,
  bins = NULL,
  binwidth = NULL,
  breaks = NULL,
  nx = 100,
  ny = 100,
  xlim = NULL,
  ylim = NULL,
  formula = z ~ 1,
  model = "Exp",
  inits = c(NA, NA, NA),
  var = FALSE,
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE
)

StatKrigeContourLines

geom_krige_contour_lines(
  mapping = NULL,
  data = NULL,
  stat = "krige_contour_lines",
  position = "identity",
  ...,
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE
)

GeomKrigeContourLines

stat_krige_contour(
  mapping = NULL,
  data = NULL,
  geom = "krige_contour",
  position = "identity",
  ...,
  bins = NULL,
  binwidth = NULL,
  breaks = NULL,
  nx = 100,
  ny = 100,
  xlim = NULL,
  ylim = NULL,
  formula = z ~ 1,
  model = "Exp",
  inits = c(NA, NA, NA),
  var = FALSE,
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE
)

StatKrigeContour

geom_krige_contour(
  mapping = NULL,
  data = NULL,
  stat = "krige_contour",
  position = "identity",
  ...,
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE
)

GeomKrigeContour
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{\link[ggplot2:aes]{aes()}} or
\code{\link[ggplot2:aes_]{aes_()}}. If specified and \code{inherit.aes = TRUE} (the
default), it is combined with the default mapping at the top level of the
plot. You must supply \code{mapping} if there is no plot mapping.}

\item{data}{The data to be displayed in this layer. There are three
options:

If \code{NULL}, the default, the data is inherited from the plot
data as specified in the call to \code{\link[ggplot2:ggplot]{ggplot()}}.

A \code{data.frame}, or other object, will override the plot
data. All objects will be fortified to produce a data frame. See
\code{\link[ggplot2:fortify]{fortify()}} for which variables will be created.

A \code{function} will be called with a single argument,
the plot data. The return value must be a \code{data.frame}, and
will be used as the layer data. A \code{function} can be created
from a \code{formula} (e.g. \code{~ head(.x, 10)}).}

\item{geom}{Use to override the default connection between \code{geom_krige()} and \code{stat_krige()}}

\item{position}{Position adjustment, either as a string, or the result of
a call to a position adjustment function.}

\item{...}{Other arguments passed on to \code{\link[ggplot2:layer]{layer()}}. These are
often aesthetics, used to set an aesthetic to a fixed value, like
\code{colour = "red"} or \code{size = 3}. They may also be parameters
to the paired geom/stat.}

\item{bins}{TODO}

\item{binwidth}{TODO}

\item{breaks}{TODO}

\item{nx}{Resolution of grid used for predictions}

\item{ny}{Resolution of grid used for predictions}

\item{xlim}{Range to compute and draw predictions. If \code{NULL}, defaults to
range of data.}

\item{ylim}{Range to compute and draw predictions. If \code{NULL}, defaults to
range of data.}

\item{formula}{TODO}

\item{model}{TODO}

\item{inits}{TODO}

\item{var}{TODO}

\item{na.rm}{If \code{FALSE}, the default, missing values are removed with
a warning. If \code{TRUE}, missing values are silently removed.}

\item{show.legend}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped.
\code{FALSE} never includes, and \code{TRUE} always includes.
It can also be a named logical vector to finely select the aesthetics to
display.}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics,
rather than combining with them. This is most useful for helper functions
that define both data and aesthetics and shouldn't inherit behaviour from
the default plot specification, e.g. \code{\link[ggplot2:borders]{borders()}}.}

\item{stat}{Use to override the default connection between \code{geom_krige()} and \code{stat_krige()}}
}
\description{
TODO
}
\section{Aesthetics}{
 geom_krige understands the following aesthetics (required
aesthetics are in bold):
\itemize{
\item \strong{x}
\item \strong{y}
\item \strong{z}
\item alpha
\item color
\item fill
\item group
\item linetype
\item size
\item subgroup
}

geom_krige_lines understands the following aesthetics (required
aesthetics are in bold):
\itemize{
\item \strong{x}
\item \strong{y}
\item \strong{z}
\item alpha
\item color
\item group
\item linetype
\item size
\item subgroup
}
}

\section{Computed variables}{


\describe{
\item{pred}{predicted value}
}
}

\keyword{datasets}
